requirements.txt
alembic==1.4.3
aniso8601==8.0.0
attrs==20.3.0
bcrypt==3.1.7
boto3==1.13.9
botocore==1.16.26
cached-property==1.5.2
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cmarkgfm==0.4.2
crypto==1.4.1
cryptography==3.3.1
dataclasses==0.7
dataset==1.3.1
distro==1.5.0
docker==4.4.0
docker-compose==1.27.4
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
Flask==1.1.2
Flask-Caching==1.8.0
Flask-Cors==3.0.9
flask-marshmallow==0.10.1
Flask-Migrate==2.5.3
flask-restx==0.2.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.3
gevent==20.9.0
greenlet==0.4.17
gunicorn==20.0.4
html5lib==1.0.1
idna==2.10
importlib-metadata==3.3.0
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
jsonschema==3.2.0
lxml==4.5.1
Mako==1.1.3
MarkupSafe==1.1.1
marshmallow==2.20.2
marshmallow-sqlalchemy==0.17.0
maxminddb==1.5.4
mysql==0.0.2
mysql-connector-python==8.0.22
mysqlclient==2.0.1
Naked==0.1.31
paramiko==2.7.2
passlib==1.7.2
protobuf==3.14.0
pyasn1==0.4.8
pycparser==2.20
pycryptodome==3.9.9
pycryptodomex==3.9.9
pydantic==1.5.1
PyMySQL==0.9.3
PyNaCl==1.4.0
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.13.0
python-editor==1.0.4
python-geoacumen==0.0.1
pytz==2020.1
PyYAML==5.3.1
redis==3.5.2
requests==2.23.0
rsa==4.6
s3transfer==0.3.3
shellescape==3.8.1
six==1.15.0
SQLAlchemy==1.3.17
SQLAlchemy-Utils==0.36.6
tenacity==6.2.0
texttable==1.6.3
typing-extensions==3.7.4.3
urllib3==1.25.10
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==1.0.1
WTForms==2.3.1
zipp==3.4.0
zope.event==4.5.0
zope.interface==5.1.2
